Strip CIF reader
Motivation: it only worked for some CIF files but not for most that one
would get from various crystal structure databases. Symmetry class just
served the purpose of representing the content of a CIF file and is
therefore removed as well. If we want this back at some point, it should
be implemented way more carefully. This started as a quick hack and
never matured much beyond that.