Add method for producing basis object compatible with gbasis
1. Add `gbasis_basis` to `IOData`
2. Add test

The method is added to IOData instead of MoleculeBasis or Shell because the
coordinates are not stored within either of these classes. Instead, the index is
stored, and we need to refer back to the IOData to use instances of these
classes. This seems to suggest that these classes will not be used outside of
the IOData class. Of course, we can move this method to the MoleculeBasis class
and pass the coordinates, but this seems to contradict the design principle
above.

Though gbasis supports spherical contractions, the implemented wrapper does not
because the current documentation ordering of spherical primitives (in
`MolecularBasis`) does not have sufficient information to do so (for me
anyways). For example, it says "Wikipedia-ordered real solid spherical
harmonics", but does not refer to any wikipedia page and the page for "real
solid spherical harmonics" does not follow the convention given. Spherical
contractions can be supported once the documentation is more clear or by the
author.

Also, pylint seems to think atcoords attribute in IOData is unsubscriptable,
which is a pain. (Error message was disabled inline)