Simplify strings in conventions for pure functions
The character for the angular momentum is omitted because it is
redundant. With this change, the conventions are no longer limited
in angular momentum by the availability of alphabetical characters.

Some mistakes in the docstring of `MolecularBasis` are also fixed.
More elaborate documentation will be added in a separate PR.