Add conversion between primitive shells and kind
The method "convert_kind" inside class MolecularBasis is added
that converts all kind of a shell to either Cartesian or Pure.
Note that the coefficients are averaged, since iodata assumes
coefficients are the same throughtout a Primitive Shell.

The function "convert_primitive_kind" allows the ability
to convert one primitive shell to another.